add serde support  to - blob - document - datetime - number
number, blob and document moves to another file
